Remove fake user agent string, only redirect on 301 response.
The reason for the user agent was that www.yelp.com would return a 303
"See Other" response directing to m.yelp.com when using the default user
agent string. In order to prevent this kind of redirect (because the
yelp app doesn't catch m.yelp.com for some reason) I faked a desktop
user agent.

This fix is to instead only respond to 301 "Moved Permanently"
redirects, which is what short url redirectors use (though maybe not in
all cases...).

Also, the MultipleRedirectEvaluatorTask now recognizes a null return
string as indicating that no redirection was required, instead of
looking for a evaluation that doesn't change the url (which makes more
sense because urls that do not redirect return no 'location' in the http
header response).

Fixes #26.